Move osme files and list CSS fixes
